+++ Cargo.toml
@@ -17,7 +17,7 @@ default-members = [".", "crates/redb-derive", "crates/redb-python"]
 
 [workspace.package]
 edition = "2024"
-rust-version = "1.88"
+rust-version = "1.89"
 license = "MIT OR Apache-2.0"
 homepage = "https://www.redb.org"
 repository = "https://github.com/cberner/redb"
@@ -27,9 +27,6 @@ authors = ["Christopher Berner <me@cberner.com>"]
 log = { version = "0.4.17", optional = true }
 chrono_v0_4 = { package = "chrono", version= "0.4.41", optional = true }
 uuid = { version= "1.17.0", optional = true }
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2.104"
 
 # Common test/bench dependencies
 [dev-dependencies]
