+++ rust-toolchain
@@ -1,1 +1,1 @@
-1.88
+1.89
